requirements.txt
astroid==2.2.5
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
mysql-connector-python==8.0.15
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
six==1.12.0
typed-ast==1.3.4
wrapt==1.11.1
